server/requirements.txt
django>=6.0
djangorestframework>=3.17
psycopg2-binary>=2.9
python-dotenv>=1.0
pytest>=9.0
pytest-django>=4.12
